+++ src/main/java/fr/chakib/houd/kata/manufacture/core/domain/drink/CoffeeDrink.java
@@ -1,4 +1,4 @@
-package fr.chakib.houd.kata.manufacture.core.domain;
+package fr.chakib.houd.kata.manufacture.core.domain.drink;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
