+++ src/main/java/fr/chakib/houd/kata/manufacture/core/domain/drink/TeaDrink.java
@@ -1,15 +1,15 @@
-package fr.chakib.houd.kata.manufacture.core.domain;
+package fr.chakib.houd.kata.manufacture.core.domain.drink;
+
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 
-import static java.lang.Float.parseFloat;
 
-public class ChocolateDrink {
+public class TeaDrink {
 
-    private static final String PROTOCOL = "H";
-    private static final BigDecimal AMOUNT = new BigDecimal("0.5");
-    private static final String INSTRUCTION = "chocolate";
+    private static final String PROTOCOL = "T";
+    private static final BigDecimal AMOUNT = new BigDecimal("0.4");
+    private static final String INSTRUCTION = "tea";
 
     public boolean validProtocols(String drinkProtocol, String amountProtocol){
         var amountReceived = new BigDecimal(amountProtocol);
@@ -22,7 +22,7 @@ public class ChocolateDrink {
         return INSTRUCTION;
     }
 
-    private BigDecimal missingAmount(BigDecimal amountProtocol){
-        return AMOUNT.subtract(amountProtocol).setScale(2, RoundingMode.HALF_UP);
+    private BigDecimal missingAmount(BigDecimal amountReceived){
+        return AMOUNT.subtract(amountReceived).setScale(2, RoundingMode.HALF_UP);
     }
 }
